feat: T67 MapScene（マップ描画シーン）・NodeType を shared/types に移動・EventBus に mapUpdate/nodeSelected 追加
