+++ src/shared/types/index.ts
@@ -34,3 +34,4 @@ export type EffectDef = {
 // --- Re-exports ---
 export type { Result } from './Result'
 export { ok, err } from './Result'
+export { NodeType } from './NodeType'
